requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.0
httpx==0.25.2
tenacity==8.2.3
anthropic==0.18.1
faiss-cpu>=1.8.0
numpy>=1.24.0
tiktoken==0.5.2
fhir.resources==7.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
